Deletition Protection on `Team & Skills DELETE`

+++ apps/admin_platform/views.py
@@ -78,7 +78,7 @@ def adminGerirEquipaProcesso_view(request):
 
 
 def adminCriarTeams_view(request):
-    
+
     criar_Team_Form = criarTeamForm(request.POST or None)
 
     if criar_Team_Form.is_valid():
@@ -93,8 +93,44 @@ def adminCriarTeams_view(request):
     return render(request, 'admin_platform/criarTeams.html', context)
 
 def adminTeams_delete_view(request, team_id):
-    get_Team_Form = Team.objects.get(id=team_id)
-    get_Team_Form.delete()
+
+    team = Team.objects.get(id=team_id)
+
+    utilizadores = UserProfile.objects.all()
+    allQueueTasks = QueueTask.objects.all()
+    allProcesses = QueueProcess.objects.all()
+    
+    teamSize = 0 
+    teamDoneProcess = False
+
+    # VERIFY IF THE TEAM HAS ONLY THE TEAMLEADER AS A MEMBER
+    for membros in utilizadores:
+        if(membros.idTeam != None and membros.idTeam.id == team.id):
+            teamSize += 1
+
+    # VERIFY IF A TASK IN PROCESS WAS STARTED OR NOT, IF YES IT WONT LET ADMIN DELETE TEAM
+    for task in allQueueTasks:
+        for process in allProcesses:
+            for membro in utilizadores:
+
+                if not (process.idUser and membro.idTeam and task.idProcess.id == process.id and task.startWorkingDate):
+                    continue
+                
+                if process.idUser.id == membro.idUser.id and membro.idTeam.id == team.id:
+                    teamDoneProcess = True
+                    break
+            
+            if teamDoneProcess:
+                break
+
+        if teamDoneProcess:
+            break
+
+
+    if(teamSize == 1 and (not teamDoneProcess)):
+        get_Team_Form = Team.objects.get(id=team_id)
+        get_Team_Form.delete()
+        
     return HttpResponseRedirect(reverse('criarTeams'))
 
 
@@ -117,6 +153,27 @@ def adminCriarSkills_view(request):
 
 
 def adminSkills_delete_view(request, skill_id):
-    get_Skill_Form = Skill.objects.get(id=skill_id)
-    get_Skill_Form.delete()
+
+    utilizadores = UserProfile.objects.all()
+    equipas = Team.objects.all()
+
+    skillInUse = False
+
+    # BOTH FOR's VERIFY IF THE SKILL WE WANT TO DELETE IS IN USE OR NOT
+
+    for equipa in equipas:
+        if equipa.idSkils.filter(id=skill_id).exists():
+            skillInUse = True
+            break
+
+    for utilizador in utilizadores:
+        if utilizador.idSkills.filter(id=skill_id).exists():
+            skillInUse = True
+            break
+
+
+    if(not skillInUse):
+        get_Skill_Form = Skill.objects.get(id=skill_id)
+        get_Skill_Form.delete()
+
     return HttpResponseRedirect(reverse('criarSkills'))
